# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Synthetic GTM Signal Engine API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /demo/generate-prospect:
    post:
      operationId: generateProspect
      tags: [demo]
      summary: Generate one synthetic prospect by archetype
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateProspectInput"
      responses:
        "201":
          description: Generated prospect detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProspectDetail"
  /demo/generate-batch:
    post:
      operationId: generateBatch
      tags: [demo]
      summary: Generate a mixed batch of synthetic prospects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateBatchInput"
      responses:
        "201":
          description: Batch generation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateBatchResult"
...
  /demo/simulate-days:
    post:
      operationId: simulateDays
      tags: [demo]
      summary: Simulate additional days of product activity for existing prospects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulateDaysInput"
      responses:
        "200":
          description: Simulation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulateDaysResult"
...
components:
  schemas:
...
    Archetype:
      type: string
      enum:
        - rapid_team_activator
        - enterprise_evaluator
        - solo_builder
        - stalled_implementer
        - returning_evaluator
        - converted_account
...
    OutreachPriority:
      type: string
      enum: [High, Medium, Low, Suppress]
    GtmSignalStatus:
      type: string
      enum:
        - Researching
        - Ready for Generation
        - Generated
        - Needs Review
        - Regeneration Requested
        - Regenerated
        - Approved
        - Rejected
        - Paused
        - Sent
        - Replied
    Company:
      type: object
      required:
        - id
        - name
        - domain
        - industry
        - employeeCount
        - employeeRange
        - fundingStage
        - latestFundingDate
        - fundingAmount
        - headquarters
        - productCategory
        - technologyContext
        - growthSignal
        - icpFitScore
        - createdAt
      properties:
        id: { type: integer }
        name: { type: string }
        domain: { type: string }
        industry: { type: array, items: { type: string } }
        employeeCount: { type: integer }
        employeeRange: { type: string }
        fundingStage: { type: string }
        latestFundingDate: { type: ["string", "null"], format: date }
        fundingAmount: { type: ["integer", "null"] }
        headquarters: { type: ["string", "null"] }
        productCategory: { type: array, items: { type: string } }
        technologyContext: { type: array, items: { type: string } }
        growthSignal: { type: string }
        icpFitScore: { type: number }
        createdAt: { type: string, format: date-time }
    Person:
      type: object
      required:
        - id
        - firstName
        - lastName
        - email
        - profileUrl
        - title
        - department
        - seniority
        - persona
        - purchaseRole
        - startDate
        - companyId
        - archetype
        - contactPriority
        - signupDate
        - lifecycleStage
      properties:
        id: { type: integer }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        profileUrl: { type: string }
        title: { type: string }
        department: { type: string }
        seniority: { type: string }
        persona: { type: string }
        purchaseRole: { type: string }
        startDate: { type: string, format: date-time }
        companyId: { type: integer }
        archetype:
          $ref: "#/components/schemas/Archetype"
        contactPriority: { type: string }
        signupDate: { type: string, format: date-time }
        lifecycleStage: { type: string }
    ProductEvent:
      type: object
      required:
        - id
        - personId
        - companyId
        - sessionId
        - eventName
        - eventCategory
        - occurredAt
        - source
        - properties
      properties:
        id: { type: integer }
        personId: { type: integer }
        companyId: { type: integer }
        sessionId: { type: string }
        eventName: { type: string }
        eventCategory: { type: string }
        occurredAt: { type: string, format: date-time }
        source: { type: string }
        properties:
          type: object
          additionalProperties: true
    BehavioralTrail:
      type: object
      required:
        - id
        - personId
        - companyId
        - chronologicalTrail
        - behaviorSummary
        - firstActivationAt
        - lastActivityAt
        - generatedAt
      properties:
        id: { type: integer }
        personId: { type: integer }
        companyId: { type: integer }
        chronologicalTrail:
          type: array
          items: { type: string }
        behaviorSummary: { type: string }
        firstActivationAt: { type: ["string", "null"], format: date-time }
        lastActivityAt: { type: ["string", "null"], format: date-time }
        generatedAt: { type: string, format: date-time }
...
    ResearchAssessment:
      type: object
      required:
        - id
        - personId
        - companyId
        - hypothesisVersion
        - icpFitScore
        - personaFitScore
        - activationScore
        - collaborationScore
        - enterpriseIntentScore
        - purchaseIntentScore
        - churnRiskScore
        - outreachPriority
        - recommendedAngle
        - rationale
        - riskNotes
        - generatedAt
      properties:
        id: { type: integer }
        personId: { type: integer }
        companyId: { type: integer }
        hypothesisVersion: { type: string }
        icpFitScore: { type: number }
        personaFitScore: { type: number }
        activationScore: { type: number }
        collaborationScore: { type: number }
        enterpriseIntentScore: { type: number }
        purchaseIntentScore: { type: number }
        churnRiskScore: { type: number }
        outreachPriority:
          $ref: "#/components/schemas/OutreachPriority"
        recommendedAngle: { type: string }
        rationale: { type: string }
        riskNotes: { type: string }
        generatedAt: { type: string, format: date-time }
    ProspectListItem:
      type: object
      required:
        - personId
        - firstName
        - lastName
        - title
        - companyId
        - companyName
        - archetype
        - icpFitScore
        - activationScore
        - enterpriseIntentScore
        - churnRiskScore
        - outreachPriority
        - outreachStatus
        - triggeredEventTypes
      properties:
        personId: { type: integer }
        firstName: { type: string }
        lastName: { type: string }
        title: { type: string }
        department: { type: string }
        companyId: { type: integer }
        companyName: { type: string }
        archetype:
          $ref: "#/components/schemas/Archetype"
        icpFitScore: { type: number }
        activationScore: { type: number }
        enterpriseIntentScore: { type: number }
        churnRiskScore: { type: number }
        outreachPriority:
          $ref: "#/components/schemas/OutreachPriority"
        outreachStatus:
          type: ["string", "null"]
        triggeredEventTypes:
          type: array
          items: { type: string }
    ProspectDetail:
      type: object
      required:
        - person
        - company
        - events
        - behavioralTrail
        - researchAssessment
        - gtmSignal
      properties:
        person:
          $ref: "#/components/schemas/Person"
        company:
          $ref: "#/components/schemas/Company"
        events:
          type: array
          items:
            $ref: "#/components/schemas/ProductEvent"
        behavioralTrail:
          $ref: "#/components/schemas/BehavioralTrail"
        researchAssessment:
          $ref: "#/components/schemas/ResearchAssessment"
        gtmSignal:
          oneOf:
            - $ref: "#/components/schemas/GtmSignal"
            - type: "null"
    GtmSignal:
      type: object
      required:
        - id
        - personId
        - companyId
        - sourceSignal
        - behavioralTrail
        - behaviorSummary
        - researchNotes
        - outreachAngle
        - hypothesisVersion
        - promptVersion
        - status
        - exportedToAttio
        - attioSyncStatus
        - attioCompanyRecordId
        - attioPersonRecordId
        - attioGtmSignalRecordId
        - attioPersonWebUrl
        - attioSyncError
        - attioSyncedAt
        - generationStatus
        - generationError
        - agentConfidence
        - outreachEmailSubject
        - outreachEmailBody
        - authProblemAngle
        - createdAt
      properties:
        id: { type: integer }
        personId: { type: integer }
        companyId: { type: integer }
        sourceSignal: { type: string }
        behavioralTrail:
          type: array
          items: { type: string }
        behaviorSummary: { type: string }
        researchNotes: { type: string }
        authProblemAngle: { type: ["string", "null"] }
        outreachAngle: { type: string }
        hypothesisVersion: { type: string }
        promptVersion: { type: string }
        status:
          $ref: "#/components/schemas/GtmSignalStatus"
        exportedToAttio: { type: boolean }
        attioSyncStatus:
          $ref: "#/components/schemas/AttioSyncStatus"
        attioCompanyRecordId: { type: ["string", "null"] }
        attioPersonRecordId: { type: ["string", "null"] }
        attioGtmSignalRecordId: { type: ["string", "null"] }
        attioPersonWebUrl: { type: ["string", "null"] }
        attioSyncError: { type: ["string", "null"] }
        attioSyncedAt: { type: ["string", "null"], format: date-time }
        generationStatus:
          $ref: "#/components/schemas/GenerationStatus"
        generationError: { type: ["string", "null"] }
        agentConfidence:
          oneOf:
            - $ref: "#/components/schemas/AgentConfidence"
            - type: "null"
        outreachEmailSubject: { type: string }
        outreachEmailBody: { type: string }
        createdAt: { type: string, format: date-time }
...
    AttioSyncStatus:
      type: string
      enum: [not_synced, synced, error]
    GenerationStatus:
      type: string
      enum: [pending, generated, failed]
    AgentConfidence:
      type: string
      enum: [low, medium, high]
...
    GenerateProspectInput:
      type: object
      required: [archetype]
      properties:
        archetype:
          $ref: "#/components/schemas/Archetype"
    GenerateBatchInput:
      type: object
      properties:
        count:
          type: integer
          minimum: 1
          maximum: 50
          default: 10
    GenerateBatchResult:
      type: object
      required: [generated, prospects]
      properties:
        generated: { type: integer }
        prospects:
          type: array
          items:
            $ref: "#/components/schemas/ProspectListItem"
    SimulateDaysInput:
      type: object
      properties:
        days:
          type: integer
          minimum: 1
          maximum: 30
          default: 7
    SimulateDaysResult:
      type: object
      required: [eventsAdded, daysSimulated, trailsUpdated]
      properties:
        eventsAdded: { type: integer }
        daysSimulated: { type: integer }
        trailsUpdated: { type: integer }