# lib/api-zod/src/generated/types/gtmSignal.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Synthetic GTM Signal Engine API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AgentConfidence } from './agentConfidence';
import type { AttioSyncStatus } from './attioSyncStatus';
import type { GenerationStatus } from './generationStatus';
import type { GtmSignalStatus } from './gtmSignalStatus';

export interface GtmSignal {
  id: number;
  personId: number;
  companyId: number;
  sourceSignal: string;
  behavioralTrail: string[];
  behaviorSummary: string;
  researchNotes: string;
  /** @nullable */
  authProblemAngle: string | null;
  outreachAngle: string;
  hypothesisVersion: string;
  promptVersion: string;
  status: GtmSignalStatus;
  exportedToAttio: boolean;
  attioSyncStatus: AttioSyncStatus;
  /** @nullable */
  attioCompanyRecordId: string | null;
  /** @nullable */
  attioPersonRecordId: string | null;
  /** @nullable */
  attioGtmSignalRecordId: string | null;
  /** @nullable */
  attioPersonWebUrl: string | null;
  /** @nullable */
  attioSyncError: string | null;
  /** @nullable */
  attioSyncedAt: Date | null;
  generationStatus: GenerationStatus;
  /** @nullable */
  generationError: string | null;
  agentConfidence: AgentConfidence | null;
  outreachEmailSubject: string;
  outreachEmailBody: string;
  createdAt: Date;
}