# lib/api-zod/src/generated/types/gtmSignalStatus.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Synthetic GTM Signal Engine API specification
 * OpenAPI spec version: 0.1.0
 */

export type GtmSignalStatus = typeof GtmSignalStatus[keyof typeof GtmSignalStatus];


export const GtmSignalStatus = {
  Researching: 'Researching',
  Ready_for_Generation: 'Ready for Generation',
  Generated: 'Generated',
  Needs_Review: 'Needs Review',
  Regeneration_Requested: 'Regeneration Requested',
  Regenerated: 'Regenerated',
  Approved: 'Approved',
  Rejected: 'Rejected',
  Paused: 'Paused',
  Sent: 'Sent',
  Replied: 'Replied',
} as const;