# lib/api-zod/src/generated/types/prospectDetail.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Synthetic GTM Signal Engine API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BehavioralTrail } from './behavioralTrail';
import type { Company } from './company';
import type { GtmSignal } from './gtmSignal';
import type { Person } from './person';
import type { ProductEvent } from './productEvent';
import type { ResearchAssessment } from './researchAssessment';

export interface ProspectDetail {
  person: Person;
  company: Company;
  events: ProductEvent[];
  behavioralTrail: BehavioralTrail;
  researchAssessment: ResearchAssessment;
  gtmSignal: GtmSignal | null;
}